+++ prettier.config.js
@@ -1,0 +1,11 @@
+/**@type {import("prettier").Config} */
+// eslint-disable-next-line no-undef
+module.exports = {
+  semi: true,
+  arrowParens: 'always',
+  endOfLine: 'lf',
+  quoteProps: 'consistent',
+  trailingComma: 'all',
+  singleQuote: true,
+  tabWidth: 2,
+};